Moved load_cal_info to managment again
