+++ imucal/__init__.py
@@ -8,7 +8,6 @@ from imucal.ferraris_calibration import (
     FerrarisSignalRegions,
 )
 from imucal.ferraris_calibration_info import FerrarisCalibrationInfo, TurntableCalibrationInfo  # noqa: F401
-from imucal.utils import load_calibration_info
 
 __all__ = [
     "CalibrationInfo",
@@ -20,5 +19,4 @@ __all__ = [
     "ferraris_regions_from_df",
     "ferraris_regions_from_interactive_plot",
     "ferraris_regions_from_section_list",
-    "load_calibration_info",
 ]
